+++ src/test/java/it/pagopa/transactions/commands/handlers/v2/TransactionActivateHandlerTest.java
@@ -1,15 +1,17 @@
-package it.pagopa.transactions.commands.handlers.v1;
+package it.pagopa.transactions.commands.handlers.v2;
 
 import io.vavr.control.Either;
 import it.pagopa.ecommerce.commons.client.QueueAsyncClient;
 import it.pagopa.ecommerce.commons.documents.BaseTransactionEvent;
 import it.pagopa.ecommerce.commons.documents.PaymentNotice;
 import it.pagopa.ecommerce.commons.documents.PaymentTransferInformation;
-import it.pagopa.ecommerce.commons.documents.v1.Transaction;
-import it.pagopa.ecommerce.commons.documents.v1.TransactionActivatedData;
-import it.pagopa.ecommerce.commons.documents.v1.TransactionActivatedEvent;
+import it.pagopa.ecommerce.commons.documents.v2.Transaction;
+import it.pagopa.ecommerce.commons.documents.v2.TransactionActivatedData;
+import it.pagopa.ecommerce.commons.documents.v2.TransactionActivatedEvent;
+import it.pagopa.ecommerce.commons.documents.v2.activation.EmptyTransactionGatewayActivationData;
+import it.pagopa.ecommerce.commons.documents.v2.activation.NpgTransactionGatewayActivationData;
 import it.pagopa.ecommerce.commons.domain.*;
-import it.pagopa.ecommerce.commons.domain.v1.TransactionEventCode;
+import it.pagopa.ecommerce.commons.domain.v2.TransactionEventCode;
 import it.pagopa.ecommerce.commons.exceptions.JWTTokenGenerationException;
 import it.pagopa.ecommerce.commons.queues.QueueEvent;
 import it.pagopa.ecommerce.commons.queues.TracingUtils;
@@ -46,13 +48,14 @@ import java.time.Duration;
 import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
-import static it.pagopa.ecommerce.commons.v1.TransactionTestUtils.*;
+import static it.pagopa.ecommerce.commons.v2.TransactionTestUtils.*;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.*;
 
 @ExtendWith(MockitoExtension.class)
-class TransactionInitializerHandlerTest {
+class TransactionActivateHandlerTest {
 
     private final PaymentRequestInfoRedisTemplateWrapper paymentRequestInfoRedisTemplateWrapper = Mockito
             .mock(PaymentRequestInfoRedisTemplateWrapper.class);
@@ -86,13 +89,17 @@ class TransactionInitializerHandlerTest {
 
     private final OpenTelemetryUtils openTelemetryUtils = Mockito.mock(OpenTelemetryUtils.class);
 
+    private static final String ORDER_ID = "orderId";
+
+    private static final UUID CORRELATION_ID = UUID.randomUUID();
+
     private static final String STRONG_KEY = "ODMzNUZBNTZENDg3NTYyREUyNDhGNDdCRUZDNzI3NDMzMzQwNTFEREZGQ0MyQzA5Mjc1RjY2NTQ1NDk5MDMxNzU5NDc0NUVFMTdDMDhGNzk4Q0Q3RENFMEJBODE1NURDREExNEY2Mzk4QzFEMTU0NTExNjUyMEExMzMwMTdDMDk";
 
     private static final int tokenValidityTimeInSeconds = 900;
 
     private final SecretKey jwtSecretKey = new SecretsConfigurations().ecommerceSigningKey(STRONG_KEY);
 
-    private final it.pagopa.transactions.commands.handlers.v1.TransactionActivateHandler handler = new TransactionActivateHandler(
+    private final TransactionActivateHandler handler = new TransactionActivateHandler(
             paymentRequestInfoRedisTemplateWrapper,
             transactionEventActivatedStoreRepository,
             nodoOperations,
@@ -109,7 +116,158 @@ class TransactionInitializerHandlerTest {
     );
 
     @Test
-    void shouldHandleCommandForNM3CachedPaymentRequest() {
+    void shouldHandleCommandForNM3CachedPaymentRequestWithNpgWithV2Api() {
+        Duration elapsedTimeFromActivation = Duration.ofSeconds(paymentTokenTimeout);
+        ZonedDateTime transactionActivatedTime = ZonedDateTime.now().minus(elapsedTimeFromActivation);
+        RptId rptId = new RptId(RPT_ID);
+        IdempotencyKey idempotencyKey = new IdempotencyKey("32009090901", "aabbccddee");
+        String paymentToken = PAYMENT_TOKEN;
+        String paName = "paName";
+        String paTaxcode = rptId.getFiscalCode();
+        TransactionId transactionId = new TransactionId(TRANSACTION_ID);
+        it.pagopa.generated.transactions.v2.server.model.NewTransactionRequestDto requestDto = new it.pagopa.generated.transactions.v2.server.model.NewTransactionRequestDto();
+        it.pagopa.generated.transactions.v2.server.model.PaymentNoticeInfoDto paymentNoticeInfoDto = new it.pagopa.generated.transactions.v2.server.model.PaymentNoticeInfoDto();
+        requestDto.addPaymentNoticesItem(paymentNoticeInfoDto);
+        paymentNoticeInfoDto.setRptId(rptId.value());
+        requestDto.setEmail(EMAIL_STRING);
+        requestDto.setOrderId(ORDER_ID);
+        paymentNoticeInfoDto.setAmount(1200);
+        TransactionActivateCommand command = new TransactionActivateCommand(
+                rptId,
+                new NewTransactionRequestData(
+                        requestDto.getIdCart(),
+                        requestDto.getEmail(),
+                        requestDto.getOrderId(),
+                        CORRELATION_ID,
+                        requestDto.getPaymentNotices().stream().map(
+                                el -> new it.pagopa.ecommerce.commons.domain.PaymentNotice(
+                                        null,
+                                        new RptId(el.getRptId()),
+                                        new TransactionAmount(el.getAmount()),
+                                        null,
+                                        null,
+                                        null,
+                                        false
+                                )
+                        ).toList()
+                ),
+                Transaction.ClientId.CHECKOUT.name(),
+                transactionId
+        );
+
+        PaymentRequestInfo paymentRequestInfoCached = new PaymentRequestInfo(
+                rptId,
+                paTaxcode,
+                paName,
+                DESCRIPTION,
+                AMOUNT,
+                dueDate,
+                paymentToken,
+                transactionActivatedTime.toString(),
+                idempotencyKey,
+                List.of(new PaymentTransferInfo(rptId.getFiscalCode(), false, AMOUNT, null)),
+                false
+        );
+
+        TransactionActivatedEvent transactionActivatedEvent = new TransactionActivatedEvent();
+        transactionActivatedEvent.setTransactionId(transactionId.value());
+        transactionActivatedEvent.setEventCode(TransactionEventCode.TRANSACTION_ACTIVATED_EVENT.toString());
+        TransactionActivatedData transactionActivatedData = new TransactionActivatedData();
+        transactionActivatedData.setPaymentNotices(
+                List.of(
+                        new PaymentNotice(
+                                paymentToken,
+                                rptId.value(),
+                                null,
+                                null,
+                                null,
+                                List.of(new PaymentTransferInformation(rptId.getFiscalCode(), false, null, null)),
+                                false
+                        )
+                )
+        );
+        transactionActivatedEvent.setData(transactionActivatedData);
+
+        /* preconditions */
+        Mockito.when(paymentRequestInfoRedisTemplateWrapper.findById(rptId.value()))
+                .thenReturn(Optional.of(paymentRequestInfoCached));
+        Mockito.when(transactionEventActivatedStoreRepository.save(any()))
+                .thenAnswer(args -> Mono.just(args.getArguments()[0]));
+        Mockito.doNothing().when(paymentRequestInfoRedisTemplateWrapper)
+                .save(paymentRequestInfoArgumentCaptor.capture());
+        Mockito.when(
+                transactionActivatedQueueAsyncClient.sendMessageWithResponse(
+                        any(QueueEvent.class),
+                        any(),
+                        durationArgumentCaptor.capture()
+                )
+        )
+                .thenReturn(Queues.QUEUE_SUCCESSFUL_RESPONSE);
+        Mockito.when(
+                jwtTokenUtils.generateToken(
+                        eq(jwtSecretKey),
+                        eq(tokenValidityTimeInSeconds),
+                        eq(new Claims(transactionId, requestDto.getOrderId(), null))
+                )
+        )
+                .thenReturn(Either.right("authToken"));
+
+        Mockito.when(confidentialMailUtils.toConfidential(EMAIL_STRING)).thenReturn(Mono.just(EMAIL));
+
+        /* run test */
+        Tuple2<Mono<BaseTransactionEvent<?>>, String> response = handler
+                .handle(command).block();
+
+        /* asserts */
+        Mockito.verify(paymentRequestInfoRedisTemplateWrapper, Mockito.times(1)).findById(rptId.value());
+        Mockito.verify(paymentRequestInfoRedisTemplateWrapper, Mockito.times(0)).save(any());
+        Mockito.verify(openTelemetryUtils, Mockito.times(1)).addSpanWithAttributes(
+                eq(OpenTelemetryUtils.REPEATED_ACTIVATION_SPAN_NAME),
+                argThat(
+                        arguments -> {
+                            String spanPaymentToken = arguments.get(
+                                    OpenTelemetryUtils.REPEATED_ACTIVATION_PAYMENT_TOKEN_ATTRIBUTE_KEY
+                            );
+                            Long spanLeftTime = arguments.get(
+                                    OpenTelemetryUtils.REPEATED_ACTIVATION_PAYMENT_TOKEN_LEFT_TIME_ATTRIBUTE_KEY
+                            );
+                            return paymentToken.equals(spanPaymentToken) && spanLeftTime != null;
+                        }
+                )
+        );
+        Mockito.verify(openTelemetryUtils, Mockito.times(0)).addErrorSpanWithException(any(), any());
+        assertNotNull(paymentRequestInfoCached.id());
+        TransactionActivatedEvent event = (TransactionActivatedEvent) response.getT1().block();
+
+        assertNotNull(event.getTransactionId());
+        assertInstanceOf(
+                NpgTransactionGatewayActivationData.class,
+                event.getData().getTransactionGatewayActivationData()
+        );
+        assertNotNull(
+                ((NpgTransactionGatewayActivationData) event.getData().getTransactionGatewayActivationData())
+                        .getOrderId()
+        );
+        assertEquals(
+                ORDER_ID,
+                ((NpgTransactionGatewayActivationData) event.getData().getTransactionGatewayActivationData())
+                        .getOrderId()
+        );
+        assertEquals(
+                CORRELATION_ID.toString(),
+                ((NpgTransactionGatewayActivationData) event.getData().getTransactionGatewayActivationData())
+                        .getCorrelationId()
+        );
+        assertNotNull(event.getEventCode());
+        assertNotNull(event.getCreationDate());
+        assertNotNull(event.getId());
+        assertEquals(paymentTokenTimeout, event.getData().getPaymentTokenValiditySeconds());
+        assertEquals(Duration.ofSeconds(transientQueueEventsTtlSeconds), durationArgumentCaptor.getValue());
+
+    }
+
+    @Test
+    void shouldHandleCommandForNM3CachedPaymentRequestWithPGS() {
         Duration elapsedTimeFromActivation = Duration.ofSeconds(paymentTokenTimeout);
         ZonedDateTime transactionActivatedTime = ZonedDateTime.now().minus(elapsedTimeFromActivation);
         RptId rptId = new RptId(RPT_ID);
@@ -201,7 +359,8 @@ class TransactionInitializerHandlerTest {
                         eq(tokenValidityTimeInSeconds),
                         eq(new Claims(transactionId, null, null))
                 )
-        ).thenReturn(Either.right("authToken"));
+        )
+                .thenReturn(Either.right("authToken"));
 
         Mockito.when(confidentialMailUtils.toConfidential(EMAIL_STRING)).thenReturn(Mono.just(EMAIL));
 
@@ -231,6 +390,10 @@ class TransactionInitializerHandlerTest {
         TransactionActivatedEvent event = (TransactionActivatedEvent) response.getT1().block();
 
         assertNotNull(event.getTransactionId());
+        assertInstanceOf(
+                EmptyTransactionGatewayActivationData.class,
+                event.getData().getTransactionGatewayActivationData()
+        );
         assertNotNull(event.getEventCode());
         assertNotNull(event.getCreationDate());
         assertNotNull(event.getId());
